+++ android/engine/src/main/java/org/smartregister/fhircore/engine/configuration/AppConfigClassification.kt
@@ -19,7 +19,8 @@ package org.smartregister.fhircore.engine.configuration
 enum class AppConfigClassification : ConfigClassification {
   APPLICATION,
   LOGIN,
-  PIN;
+  PIN,
+  SYNC;
 
   override val classification: String = name.lowercase()
 }