+++ android/engine/src/test/java/org/smartregister/fhircore/engine/ui/pin/PinViewModelTest.kt
@@ -32,10 +32,11 @@ import org.junit.Before
 import org.junit.Ignore
 import org.junit.Rule
 import org.junit.Test
-import org.smartregister.fhircore.engine.auth.AccountAuthenticator
-import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry
+import org.smartregister.fhircore.engine.app.fakes.Faker
+import org.smartregister.fhircore.engine.auth.AuthCredentials
 import org.smartregister.fhircore.engine.configuration.view.PinViewConfiguration
 import org.smartregister.fhircore.engine.configuration.view.pinViewConfigurationOf
+import org.smartregister.fhircore.engine.data.local.DefaultRepository
 import org.smartregister.fhircore.engine.robolectric.RobolectricTest
 import org.smartregister.fhircore.engine.rule.CoroutineTestRule
 import org.smartregister.fhircore.engine.util.DispatcherProvider
@@ -58,9 +59,8 @@ internal class PinViewModelTest : RobolectricTest() {
   @BindValue val sharedPreferencesHelper: SharedPreferencesHelper = mockk()
   @BindValue val secureSharedPreference: SecureSharedPreference = mockk()
 
-  @Inject lateinit var accountAuthenticator: AccountAuthenticator
-  @Inject lateinit var configurationRegistry: ConfigurationRegistry
-
+  val defaultRepository: DefaultRepository = mockk()
+  @BindValue var configurationRegistry = Faker.buildTestConfigurationRegistry(defaultRepository)
   private val application = ApplicationProvider.getApplicationContext<Application>()
 
   private lateinit var pinViewModel: PinViewModel
@@ -68,7 +68,7 @@ internal class PinViewModelTest : RobolectricTest() {
   private val testPin = MutableLiveData("1234")
   val testPinViewConfiguration =
     PinViewConfiguration(
-      appId = "ancApp",
+      appId = "appId",
       classification = "classification",
       applicationName = "Test App",
       appLogoIconResourceFile = "ic_launcher",
@@ -80,14 +80,20 @@ internal class PinViewModelTest : RobolectricTest() {
   fun setUp() {
     hiltRule.inject()
 
-    configurationRegistry.loadAppConfigurations("appId", accountAuthenticator) {}
-
     coEvery { sharedPreferencesHelper.read(any(), "") } returns "1234"
     coEvery { sharedPreferencesHelper.write(FORCE_LOGIN_VIA_USERNAME, true) } returns Unit
     coEvery { sharedPreferencesHelper.remove(any()) } returns Unit
     coEvery { secureSharedPreference.retrieveSessionUsername() } returns "demo"
     coEvery { secureSharedPreference.saveSessionPin("1234") } returns Unit
     coEvery { secureSharedPreference.retrieveSessionPin() } returns "1234"
+    coEvery {
+      secureSharedPreference.saveCredentials(
+        AuthCredentials("username", "password", "sessionToken", "refreshToken")
+      )
+    } returns Unit
+    coEvery { secureSharedPreference.retrievePinCredentials() } returns
+      AuthCredentials("username", "password", "sessionToken", "refreshToken")
+    coEvery { secureSharedPreference.savePinCredentials() } returns Unit
 
     pinViewModel =
       PinViewModel(
@@ -111,7 +117,7 @@ internal class PinViewModelTest : RobolectricTest() {
   fun testPinViewConfiguration() {
     val expectedPinConfig =
       pinViewConfigurationOf(
-        appId = "ancApp",
+        appId = "appId",
         classification = "classification",
         applicationName = "Test App",
         appLogoIconResourceFile = "ic_launcher",