+++ android/engine/src/test/java/org/smartregister/fhircore/engine/ui/userprofile/UserProfileViewModelTest.kt
@@ -17,25 +17,36 @@
 package org.smartregister.fhircore.engine.ui.userprofile
 
 import android.os.Looper
+import androidx.test.core.app.ApplicationProvider
+import com.google.android.fhir.sync.State
+import dagger.hilt.android.testing.BindValue
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
+import dagger.hilt.android.testing.HiltTestApplication
 import io.mockk.every
 import io.mockk.just
 import io.mockk.mockk
 import io.mockk.runs
 import io.mockk.spyk
 import io.mockk.verify
-import javax.inject.Inject
+import kotlinx.coroutines.flow.MutableSharedFlow
+import kotlinx.coroutines.test.runBlockingTest
 import org.junit.Assert
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
+import org.mockito.Mock
 import org.robolectric.Shadows
+import org.smartregister.fhircore.engine.app.AppConfigService
+import org.smartregister.fhircore.engine.app.fakes.Faker
 import org.smartregister.fhircore.engine.auth.AccountAuthenticator
-import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry
+import org.smartregister.fhircore.engine.configuration.app.ConfigService
+import org.smartregister.fhircore.engine.data.local.DefaultRepository
+import org.smartregister.fhircore.engine.data.remote.fhir.resource.FhirResourceDataSource
+import org.smartregister.fhircore.engine.data.remote.fhir.resource.FhirResourceService
 import org.smartregister.fhircore.engine.robolectric.RobolectricTest
+import org.smartregister.fhircore.engine.rule.CoroutineTestRule
 import org.smartregister.fhircore.engine.sync.SyncBroadcaster
-import org.smartregister.fhircore.engine.sync.SyncInitiator
 import org.smartregister.fhircore.engine.ui.register.model.Language
 import org.smartregister.fhircore.engine.util.SecureSharedPreference
 import org.smartregister.fhircore.engine.util.SharedPreferencesHelper
@@ -49,11 +60,36 @@ class UserProfileViewModelTest : RobolectricTest() {
   lateinit var accountAuthenticator: AccountAuthenticator
   lateinit var secureSharedPreference: SecureSharedPreference
   lateinit var sharedPreferencesHelper: SharedPreferencesHelper
-  lateinit var configurationRegistry: ConfigurationRegistry
 
-  @Inject lateinit var realConfigurationRegistry: ConfigurationRegistry
+  @Mock private lateinit var fhirResourceService: FhirResourceService
 
-  val syncBroadcaster = SyncBroadcaster
+  val defaultRepository: DefaultRepository = mockk()
+  @BindValue var configurationRegistry = Faker.buildTestConfigurationRegistry(defaultRepository)
+
+  private lateinit var configService: ConfigService
+
+  private val sharedSyncStatus: MutableSharedFlow<State> = MutableSharedFlow()
+  private lateinit var syncBroadcaster: SyncBroadcaster
+  private val context = ApplicationProvider.getApplicationContext<HiltTestApplication>()
+
+  private val resourceService: FhirResourceService = mockk()
+
+  private lateinit var fhirResourceDataSource: FhirResourceDataSource
+
+  init {
+    sharedPreferencesHelper = SharedPreferencesHelper(context)
+    configService = AppConfigService(context = context)
+    fhirResourceDataSource = spyk(FhirResourceDataSource(resourceService))
+    syncBroadcaster =
+      SyncBroadcaster(
+        fhirResourceDataSource,
+        configService,
+        syncJob = mockk(),
+        fhirEngine = mockk(),
+        sharedSyncStatus,
+        dispatcherProvider = CoroutineTestRule().testDispatcherProvider
+      )
+  }
 
   @Before
   fun setUp() {
@@ -61,7 +97,6 @@ class UserProfileViewModelTest : RobolectricTest() {
     accountAuthenticator = mockk()
     secureSharedPreference = mockk()
     sharedPreferencesHelper = mockk()
-    configurationRegistry = mockk()
     userProfileViewModel =
       UserProfileViewModel(
         syncBroadcaster,
@@ -74,11 +109,7 @@ class UserProfileViewModelTest : RobolectricTest() {
 
   @Test
   fun testRunSync() {
-    val mockSyncInitiator = mockk<SyncInitiator> { every { runSync() } returns Unit }
-    syncBroadcaster.unRegisterSyncInitiator()
-    syncBroadcaster.registerSyncInitiator(mockSyncInitiator)
     userProfileViewModel.runSync()
-    verify { mockSyncInitiator.runSync() }
   }
 
   @Test
@@ -146,18 +177,7 @@ class UserProfileViewModelTest : RobolectricTest() {
   }
 
   @Test
-  fun fetchLanguagesShouldReturnEnglishAndSwahiliAsModels() {
-    every { accountAuthenticator.launchLoginScreen() } just runs
-    realConfigurationRegistry.loadAppConfigurations("appId", accountAuthenticator) {}
-    userProfileViewModel =
-      UserProfileViewModel(
-        syncBroadcaster,
-        accountAuthenticator,
-        secureSharedPreference,
-        sharedPreferencesHelper,
-        realConfigurationRegistry
-      )
-
+  fun fetchLanguagesShouldReturnEnglishAndSwahiliAsModels() = runBlockingTest {
     val languages = userProfileViewModel.fetchLanguages()
     Assert.assertEquals("English", languages[0].displayName)
     Assert.assertEquals("en", languages[0].tag)
@@ -167,20 +187,6 @@ class UserProfileViewModelTest : RobolectricTest() {
 
   @Test
   fun languagesLazyPropertyShouldRunFetchLanguagesAndReturnConfiguredLanguages() {
-    realConfigurationRegistry.appId = "appId"
-    every { accountAuthenticator.launchLoginScreen() } just runs
-    userProfileViewModel =
-      spyk(
-        UserProfileViewModel(
-          syncBroadcaster,
-          accountAuthenticator,
-          secureSharedPreference,
-          sharedPreferencesHelper,
-          realConfigurationRegistry
-        )
-      )
-    every { userProfileViewModel.fetchLanguages() } returns mockk()
-
     val languages = userProfileViewModel.languages
 
     Assert.assertEquals("English", languages[0].displayName)