+++ android/quest/src/main/java/org/smartregister/fhircore/quest/configuration/view/ResultDetailsNavigationConfiguration.kt
@@ -20,16 +20,17 @@ import androidx.compose.runtime.Stable
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import org.smartregister.fhircore.engine.configuration.Configuration
+import org.smartregister.fhircore.engine.configuration.view.NavigationAction
+import org.smartregister.fhircore.engine.configuration.view.NavigationOption
 
 @Serializable
 @SerialName("test_details")
-data class TestDetailsNavigationAction(val form: String, val readOnly: Boolean) :
-  NavigationAction()
+data class TestDetailsNavigationAction(val form: String, val readOnly: Boolean) : NavigationAction
 
 @Stable
 @Serializable
 data class ResultDetailsNavigationConfiguration(
-  override val appId: String,
-  override val classification: String,
-  val navigationOptions: List<NavigationOption>
+  override val appId: String = "",
+  override val classification: String = "",
+  val navigationOptions: List<NavigationOption> = listOf()
 ) : Configuration