+++ CHANGELOG.md
@@ -53,6 +53,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - ANC | Resolved a bug on measure reporting : App crash when All patient selected
 - ANC | Resolved a bug in the Patient details : App crash when computing BMI
+- ANC | Refactor measure-report evaluate
 
 - EIR | Resolved a bug in the vaccination flow : App crush when saving vaccine
 - EIR | Ability for a complete immunixation flow from Record 1st Vaccine to Full Immunization
@@ -64,15 +65,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Quest | Fix patient registration with estimated age/dob
 - Quest | Fix patient register data list items display
 - Quest | Fix and reduce the time taken to load patient details page
+- Quest | Fix MLS for questionnaire and questionnaire response titles on the patient profile page
 
 - Engine | Fixed app crash when wrong appId is provided
 - Engine | Fixed login error message
+- Engine | Fixed redirect after logout when press back button
+- Engine | Fixed MLS in forms for Android 6(API 23) and below
+- Engine | Fixed logout when sync failed
 
 ### Changed
 - EIR | Updated overdue trigger flow for Vaccine Due date
 - EIR | Updated on save vaccine received alert dialogue
 
 - ANC | Removed custom extraction handling and moved Family/ANC forms to structure map
+- ANC | Implement CHW ANC BMI questionnaire extraction using StructureMap
 
 - Quest | Updated Sync to sync user specific fhir resources using Search Parameters from assets
 - Quest | Updated Patient Details View Configurations and added configurations for routing to Test details page